requirements.txt
Django==1.10.4 # Updated from 1.9.7
django-allauth==0.29.0 # Updated from 0.25.2
djangorestframework==3.5.3 # Updated from 3.3.3
djangorestframework-jwt==1.9.0 # Updated from 1.7.2
django-rest-auth==0.8.2
django-rest-swagger==2.1.0
dry-rest-permissions==0.1.8
django-cors-middleware==1.3.1
django-import-export==0.5.1
django-filter==1.0.1 # Updated from 0.13.0
gunicorn==19.6.0 # Updated from 19.4.5
invoke==0.14.0 # Updated from 0.13.0
psycopg2==2.6.2 # Updated from 2.6.1
daphne==1.0.3
asgi-redis==1.0.0
channels==1.0.2
